templating: added
the templating package is the standard Meteor Template package overridden to support Ember handlebars templates.

Any file with the extension .html .hbs .handlebars that is places in a directory called "templates" will automatically be precompiled and added to Ember.TEMPLATES. The location of the file minus the template prefix is use for the template name. For example a file in 'client/templates/todos/index.hbs' will be name "todos/index".