Replaced node-jquery with local version.
node-jQuery was not working when deploying to meteor.com. Persumily due to the use of binaries from jsdom.

Importing the packages. jQuery, Handelbars and Ember using the path 'packages/**/*.js' is not a good long term solution. In the future the library should be imported using the Meteor build system.